// src/infrastructure/grpc/generated/chat_service.ts
export interface ChatData {
  id: string;
  studentId: string;
  instructorId: string;
  createdAt: string;
  updatedAt: string;
  lastMessageId: string;
  /** viewer specific */
  isPinned: boolean;
  isArchived: boolean;
  isMuted: boolean;
}
// ...
export const ChatData: MessageFns<ChatData> = {
  encode(
    message: ChatData,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.id !== '') {
      writer.uint32(10).string(message.id);
    }
    if (message.studentId !== '') {
      writer.uint32(26).string(message.studentId);
    }
    if (message.instructorId !== '') {
      writer.uint32(34).string(message.instructorId);
    }
    if (message.createdAt !== '') {
      writer.uint32(42).string(message.createdAt);
    }
    if (message.updatedAt !== '') {
      writer.uint32(50).string(message.updatedAt);
    }
    if (message.lastMessageId !== '') {
      writer.uint32(58).string(message.lastMessageId);
    }
    if (message.isPinned !== false) {
      writer.uint32(160).bool(message.isPinned);
    }
    if (message.isArchived !== false) {
      writer.uint32(168).bool(message.isArchived);
    }
    if (message.isMuted !== false) {
      writer.uint32(176).bool(message.isMuted);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ChatData {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChatData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.studentId = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.instructorId = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.createdAt = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.updatedAt = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.lastMessageId = reader.string();
          continue;
        }
        case 20: {
          if (tag !== 160) {
            break;
          }

          message.isPinned = reader.bool();
          continue;
        }
        case 21: {
          if (tag !== 168) {
            break;
          }

          message.isArchived = reader.bool();
          continue;
        }
        case 22: {
          if (tag !== 176) {
            break;
          }

          message.isMuted = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },
};
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
}